feat(server): chart generation

+++ server/src/main/java/domain/Statistic.java
@@ -1,4 +1,4 @@
 package domain;
 
-public record Statistic(long time, int variant, int threads) {
+public record Statistic(int time, int variant, int threads) {
 }
